YEY! From bilbo-macos!! ca2 (resident, waven, ...), bilbo and Mummi contribution!! veriverse contribution!! LCTV contribution!! catherinacortega, chubz, doraemon (currently... also gosu - hiiamgosu, dum - noticemesenpaiiamdum...) fuel (Benjamin Kuhl), r00tkiller (Alexandre), epryka (Vova - Vladimir)), theaifam5 (Mateus), vapaamies (Vladislav), pablosj (Pawel Kulesza), hiiamgosu, phanxgames, intelijens and zzzel (Marko Zel) contribution!! danielv, dzetko, dopedopedope, georgeb, johnyblaze, kaeris, kranalex and kissloryshy contribution!! dzhekson6000, karrmarr, matroskinb and michaelphp contribution!! bigboom, codeotea, dopedopedope, intelijens, jilizart, jookeer, marianyc, mushmouth, peter_cados, smart6307 (thunder will/canal2048), sonnystarminecraft and vapaamies contribution!! fauker, hiiamgosu (Gosu, litAsHellhahah) and marianyc contribution!! Twitch contribution!! Gothiccat1 contribution!! Beam.pro contribution!!
Thank you Carlos!!

Thank you Mummi!!

from bilbo-OS X!!
i7 6700 32 GB RAM XPS 8900 macOS Sierra 10.12

+++ appseed/axis/net/sockets/bsd/basic/sockets_tcp_socket.cpp
@@ -1141,7 +1141,7 @@ namespace sockets
                   && x509_err != X509_V_ERR_SELF_SIGNED_CERT_IN_CHAIN
                   && x509_err != X509_V_ERR_UNABLE_TO_GET_ISSUER_CERT_LOCALLY)
             {
-               log("SSLNegotiate/cert_common_name_check",0,"cert_common_name_check failed (error="+::str::from(x509_err)+",url=" + m_strUrl + ")",::aura::log::level_info);
+               log("SSLNegotiate/cert_common_name_check",0,"cert_common_name_check failed (error="+::str::from((int)x509_err)+",url=" + m_strUrl + ")",::aura::log::level_info);
                SetSSLNegotiate(false);
                SetCloseAndDelete();
                OnSSLConnectFailed();
